change hierarchy and names, add pyproject.toml for easier black execution. remove tf.stft-related notebook. update version in __init__.py.
