+++ kapre/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.1.7'
+__version__ = '0.1.8'
 VERSION = __version__
 
 from . import time_frequency
