+++ pyproject.toml
@@ -1,0 +1,24 @@
+[tool.black]
+line-length = 100
+target-version = ['py37']
+skip-string-normalization = true
+include = '\.pyi?$'
+exclude = '''
+
+(
+  /(
+      \.eggs         # exclude a few common directories in the
+    | \.git          # root of the project
+    | \.hg
+    | \.mypy_cache
+    | \.tox
+    | \.venv
+    | _build
+    | buck-out
+    | build
+    | dist
+  )/
+  | exclude_example_foo.py # also separately exclude a file named foo.py in
+                           # the root of the project
+)
+'''
